vod management service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt>=4.0.0
pyjwt
requests
beautifulsoup4